pyproject.toml
[tool.poetry]
name = "sellit_api"
version = "0.1.0"
description = "MTTA school project"
authors = ["Adam Jankinc", "Filip Brat"]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4"
python-dotenv = "^0"
argon2-cffi = "^21.0.0"
sentry-sdk = "^1"
psycopg2-binary = "^2.8.6"
porcupine-python = "^0"
django-imap-backend = "^0"
django-celery-beat = "^2.2.0"
redis = "^4.0"
django_api_forms = "1.0.0rc7"
django-filter = "^22.1"
django-object-checker = "^1.0.2"

[tool.poetry.dev-dependencies]
fabric = "^2.6.0"
flake8 = "^3.8.4"

[build-system]
requires = ["poetry>=1.0"]
build-backend = "poetry.masonry.api"
